Tests/LibWeb/Text/expected/css/abspos-only-child-of-span-geometry.txt

target rect: 0,10 10x10
elementFromPoint hits target: true
host rect count: 1
